% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_signal.R
\name{get_average_signal}
\alias{get_average_signal}
\title{Get a dataframe of average signal from bw within regions}
\usage{
get_average_signal(bw, regions, by = NULL)
}
\arguments{
\item{bw}{}

\item{regions}{}

\item{by}{}
}
\value{

}
\description{
Get a dataframe of average signal from bw within regions
}
