% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bw_matrix.R
\name{get_bw_matrix}
\alias{get_bw_matrix}
\title{Get signal matrix for regions using bigwig input}
\usage{
get_bw_matrix(
  bw,
  regions,
  type = c("mean", "min", "max", "coverage", "sd"),
  by = NULL
)
}
\arguments{
\item{bw}{an [rtracklayer::BigWigFile], [rtracklayer::BigWigFileList] or path to bigwig file}

\item{regions}{a GRanges object (all regions must be equal width)}

\item{type}{one of: c("mean", "min", "max", "coverage", "sd")}
}
\value{
a matrix of bigwig signal.
}
\description{
Get signal matrix for regions using bigwig input
}
